// src/sanity.types.ts
export type SITE_SETTINGS_QUERY_RESULT = {
  navigation: null;
  socialLinks: null;
} | {
  navigation: Array<{
    label: string | null;
    href: string | null;
  }> | null;
  socialLinks: Array<{
    platform: "facebook" | "instagram" | "other" | "tiktok" | "youtube" | null;
    url: string | null;
  }> | null;
} | null;